+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Fixed an issue which caused a crash when all invisible instances was destroyed.
+
 ## [1.6.1] - 2022-08-19
 ### Fixed
 - Fixed an issue which caused glTF loading function to return before all resources was loaded.